Make ellipsoids instances not subclasses (#9)
It's awkward to have to use `WGS84()` when all ellipsoids don't really
take any arguments. Make them instances of the `Ellipsoid` class instead
of subclasses to get rid of `()`. Move the list of ellipsoids to a
separate documentation page and move doctests to that page. Add required
pytest configuration to the Makefile so we can run the doctests in the
rst sources.