Add a method for the ellipsoid geocentric radius (#37)
This is the distance from the center of the ellipsoid to its surface as
a function of geodetic or geocentric latitude. We usually avoid doing
conversions inside functions but this had to be an exception. It 
simply wouldn't be an option here since there is no way to do the 
conversion without knowing the radius already.

Fixes #34