Fix missing centrifugal term for sphere at the equator (#48)
Fix the method that computes the norm of the gravity acceleration vector
on the equator of a sphere. The method used to return only the
gravitational attraction term, without the centrifugal acceleration
term. Add the additional term and update the test function, which now
compares gravity on poles and equator with the normal gravity formula.