Fix normal gravity equation of Sphere (#52)
The definition we're using for normal gravity is the "magnitude of the
normal gravity potential", which is not what was being calculated for
the sphere (sum of the magnitudes of the centrifugal and gravitational
accelerations). Fix the equation so that both Sphere and Ellipsoid 
have same concept of normal gravity.