Add checks for ellipsoid parameters (#45)
Add methods for validating ellipsoid parameters. Raise error if flattening is
equal to zero, but raise warning if it's lower than 1e-7. Raise error if
flattening is negative or greater than 1. Raise error if semimajor axis is not
positive. Raise warning if GM constant is negative. Add tests functions for new
features.