Specify oldest supported version of dependencies (#111)
Following the guidance in fatiando/community#40 and NEP29, support all
minor versions from the past 24 months. This means numpy >= 1.19
and attrs >= 19.3. Add CI tests for the oldest versions but test only
oldest supported versions with the oldest Python and vice-versa because
numpy doesn't supply wheels for old versions on new Python. It also
saves CI compute time and should be more than enough of a safeguard.