Use jupyter-sphinx and update front page (#112)
Modernize the front page and installing guide to match the latest 
version of Pooch. Uses sphinx-design instead since it's meant to
replace sphinx-panels in the future. Move the tutorials to 
jupyer-sphinx instead of sphinx-gallery since we're not building 
galleries anyway. Added a glossary that can be populated with 
definitions afterwards.