Refactor the documentation of the Ellipsoid class (#127)
Add more docstring info for each derived attribute. Standardize the
maths and unit representations. Add more context to the methods docs,
including a citation to Lakshmanan (1991) for the original derivation of
general normal gravity. Includes some tweaks to the autosummary template
for classes and slight re-definitions of derived attributes to be based
on the flattening when possible.