Refactor the documentation tutorials (#130)
Shifted the focus to how to do things with Boule and point to other pages (like
the new glossary and docstrings) for more of the theory. Gravity disturbance
tutorial expanded a bit to be more in line with the Harmonica tutorial.
Coordinate conversion tutorial was refocused on using pymap3d instead of
explaining what the coordinate systems are, which should be in a separate page.