Add true mean radius for Sphere, Ellipsoid and TriaxialEllipsoid (#177)
Add a property that computes the "true" mean radius of the ellipsoid 
(the integral of the geocentric surface radius). Rename the existing 
`mean_radius` to `semiaxes_mean_radius`. These are not too different
numerically for the Earth so it doesn't make much of a difference that
the value will change. The new `mean_radius` computes the integral of
the radius using Gauss Legendre quadrature. 

There is a single parameter that needs to be defined, which is the
number of quadrature points to use. (For triaxial ellipsoids, I set the
number of points in longitude to be twice that in latitude). For the
ellipsoids that are already defined in boule, tests show that about n=10
is sufficient to get machine precision results. Tests for extreme
ellipsoids with a flattening of 0.5 (which is 100 times greater than for
Mars) requires n=30. Tests for a Vesta-like object where the semimajor
axis is reduced by a factor of two requires n=38. In an extreme
abundance of caution, I have chosen to use n=50, which is probably
sufficient for any solar system object. I have cross checked these
results using the degree-0 spherical harmonic coefficients calculated
with the independent code pyshtools, which compare to the last digit (or
two).