Add area and area_equivalent_radius to all three ellipsoid classes (#178)
Add the properties `area` and `area_equivalent_radius` to the three 
ellipsoid classes. The "area_equivalent_radius" is what WGS84
calls `R_2`.

The equations for the area of bi-axial and tri-axial ellipsoids are
taken from https://en.wikipedia.org/wiki/Ellipsoid#Surface_area. The
equation for a bi-axial ellispoid is relatively simple, and only
requires to use `np.arctanh`. The equation for a triaxial ellipsoid,
however, requires computing an elliptic integral. To do this, I have
used the elliptic integral function from scipy: `scipy.special.elliprg`. 
This required adding scipy>=1.8.0 as a dependency.

I have verified that the area equivalent radius for WGS84 in boule is
what is given in the literature. I have verified that the triaxial
ellipsoid routine gives the correct value when a=b for WGS84.

---------

Co-authored-by: Leonardo Uieda <leo@uieda.com>