Add new planetary ellipsoids and new naming scheme (#180)
Add ellipsoids for several solar system bodies. Names are now 
CamelCase and include the year at the end so we can have multiple
of the same body. 

Add the optional `comments` attribute to all classes to provide additional
 information for understanding the origin of the numbers used to 
instantiate the class (for example, many papers provide several different 
versions of the ellipsoid, and you need to choose one).

Add the `TriaxialEllipsoid.semimajor_axis_longitude` optional attribute. 
Previously, this parameter was found as an optional argument to the 
`geocentric radius` method but it's an intrinsic property of the ellipsoid 
and the coordinate frame, and will be necessary for all routines that make 
use of longitude coordinates. 

The Vesta ellipsoid was replaced since the there is not enough information 
in the paper to determine the GM, it is not stated if the ellipsoid is 
geocentric or offset from the center of mass, and the directions of the 
principal axes are not given. Instead, two new ellipsoids have been added: 
A biaxial `Vesta2017` and `VestaTriaxial2017`. The triaxial one is not aligned 
with the topography and gravity coordinate axes, and this is specified by 
the new `semimajor_axis_longitude` attribute.

---------

Co-authored-by: Leonardo Uieda <leo@uieda.com>