Replace "geodetic" argument in geocentric_radius
Use `coordinate_system` instead since we will be moving towards that
definition in other functions (see #175). Create a function to check
that the argument has a valid value. This way all arguments relating to
coordinate systems will be consistent and we can add support for
ellipsoidal harmonic coordinates as well.